requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
requests-html>=0.10.0
lxml>=4.9.0
websockets>=10.0,<11.0
selenium
webdriver-manager
google-genai
lxml[html_clean]
nest_asyncio
playwright

